# openapi/openapi.yaml
components:
  schemas:
...
    Cookie:
      additionalProperties: false
      properties:
        Domain:
          type: string
        Expires:
          format: date-time
          type: string
        HttpOnly:
          type: boolean
        MaxAge:
          format: int64
          type: integer
        Name:
          type: string
        Partitioned:
          type: boolean
        Path:
          type: string
        Quoted:
          type: boolean
        Raw:
          type: string
        RawExpires:
          type: string
        SameSite:
          format: int64
          type: integer
        Secure:
          type: boolean
        Unparsed:
          items:
            type: string
          type:
            - array
            - "null"
        Value:
          type: string
      required:
        - Name
        - Value
        - Quoted
        - Path
        - Domain
        - Expires
        - RawExpires
        - MaxAge
        - Secure
        - HttpOnly
        - SameSite
        - Partitioned
        - Raw
        - Unparsed
      type: object
...
    ErrorDetail:
      additionalProperties: false
      properties:
        location:
          description: Where the error occurred, e.g. 'body.items[3].tags' or 'path.thing-id'
          type: string
        message:
          description: Error message text
          type: string
        value:
          description: The value at the given location
      type: object
    ErrorModel:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - https://example.com/schemas/ErrorModel.json
          format: uri
          readOnly: true
          type: string
        detail:
          description: A human-readable explanation specific to this occurrence of the problem.
          examples:
            - Property foo is required but is missing.
          type: string
        errors:
          description: Optional list of individual error details
          items:
            $ref: "#/components/schemas/ErrorDetail"
          type:
            - array
            - "null"
        instance:
          description: A URI reference that identifies the specific occurrence of the problem.
          examples:
            - https://example.com/error-log/abc123
          format: uri
          type: string
        status:
          description: HTTP status code
          examples:
            - 400
          format: int64
          type: integer
        title:
          description: A short, human-readable summary of the problem type. This value should not change between occurrences of the error.
          examples:
            - Bad Request
          type: string
        type:
          default: about:blank
          description: A URI reference to human-readable documentation for the error.
          examples:
            - https://example.com/errors/example
          format: uri
          type: string
      type: object
...
    FileListOutputBody:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - https://example.com/schemas/FileListOutputBody.json
          format: uri
          readOnly: true
          type: string
        items:
          items:
            $ref: "#/components/schemas/FileObjectBody"
          type:
            - array
            - "null"
      required:
        - items
      type: object
    FileObjectBody:
      additionalProperties: false
      properties:
        attachedToId:
          format: uuid
          type: string
        attachedToType:
          examples:
            - customer_signal
          type: string
        byteSize:
          format: int64
          type: integer
        contentType:
          examples:
            - text/plain
          type: string
        createdAt:
          format: date-time
          type: string
        originalFilename:
          type: string
        publicId:
          format: uuid
          type: string
        purpose:
          examples:
            - attachment
          type: string
        sha256Hex:
          type: string
        status:
          examples:
            - active
          type: string
        tenantId:
          format: int64
          type: integer
        updatedAt:
          format: date-time
          type: string
      required:
        - publicId
        - tenantId
        - purpose
        - originalFilename
        - contentType
        - byteSize
        - sha256Hex
        - status
        - createdAt
        - updatedAt
      type: object
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http
    cookieAuth:
      in: cookie
      name: SESSION_ID
      type: apiKey
    m2mBearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http
info:
  title: HaoHao API
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /api/v1/files:
    get:
      operationId: listFiles
      parameters:
        - in: cookie
          name: SESSION_ID
          schema:
            $ref: "#/components/schemas/Cookie"
        - explode: false
          in: query
          name: attachedToType
          schema:
            type: string
        - explode: false
          in: query
          name: attachedToId
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileListOutputBody"
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      security:
        - cookieAuth: []
      summary: active tenant の file metadata を返す
      tags:
        - files
  /api/v1/files/{filePublicId}:
    delete:
      operationId: deleteFile
      parameters:
        - in: cookie
          name: SESSION_ID
          schema:
            $ref: "#/components/schemas/Cookie"
        - in: header
          name: X-CSRF-Token
          required: true
          schema:
            type: string
        - in: path
          name: filePublicId
          required: true
          schema:
            format: uuid
            type: string
      responses:
        "204":
          description: No Content
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      security:
        - cookieAuth: []
      summary: active tenant の file を soft delete する
      tags:
        - files
...